[K/N] Check that no native bridges generated with `-Xccall-mode direct`
77cfd8c disables compiling of the generated C/Objective-C source file
when `-Xccall-mode direct` is used.
It therefore disables putting this code as LLVM bitcode into cinterop
klibs.

But cinterop might still generate a non-empty C/Objective-C source file
in this mode, so this file is just ignored.
The previous commit improves that by removing one more piece of this
file -- generated protocol getters.

The only thing left in the generated file is the "preamble lines"
(header includes and code specified after `---` in the `.def` file).
Normally those lines shouldn't produce any meaningful code consumed by
the K/N compiler (when they do, it is likely a mistake).

This commit adjusts cinterop accordingly: the preamble lines are not
added to the file anymore (it is ignored anyway), and everything else is
now checked to be absent.

^KT-79751