~~ bring the number of failed tests down from 86 to 60
The idea is that if the first condition fails, the whole axis is now 0,
and doesn't participate in the comparison whatsoever, and if it's one,
then we can still choose between "captured" and "declared" by dropping
it to 0. So, if we "branch off in a linear fashion," we can employ this
trick to compactify multiple axi into a single one.