K2: fix isLocal usages in FirDeclarationsResolveTransformer broken by 17fa9d94
Before 17fa9d94, FirDeclarationsResolveTransformer contained a
FirVariable.isLocal extension that was used a number of times.
In 17fa9d94, a member FirCallableDeclaration.isLocal was introduced;
this member supersedes the extension. The extension was renamed,
but not all usages were renamed with it. In this commit,
I renamed all remaining usages.