[FIR] Fix KT-69326
Change the way `hasOnlyIncorporatedConstraintsFromDeclaredUpperBound()`
is calculated.

This change is important because it replaces a pretty strange ad-hoc
relation between `HAS_NO_DEPENDENCIES_TO_OTHER_VARIABLES` and
`HAS_CONSTRAINTS_OTHER_THAN_INCORPORATED_FROM_DECLARED_UPPER_BOUND` with
a proper condition. Not sure if it's possible to simplify the code
further without it...

^KT-69326 Fixed
^KT-77939