[K/N] Extract `CCallMode.select` to reuse it for protocol type checks
KT-79741 aims to provide an alternative, "direct" implementation for
type checks against Objective-C protocols in the compiler.
The compiler is supposed to select between the old and the new
implementation based on the existing `cCallMode` binary option.

The same selection logic is already present for function calls and
global accesses, and this commit extracts it to a new generic function,
`CCallMode.select`, to be reused for protocol type checks generation.

^KT-79751