[K/N] Add `.binaryName` to `interop.indexer.ObjCProtocol`
KT-79741 aims to provide an alternative, "direct" implementation for
type checks against Objective-C protocols in the compiler.
To achieve that, cinterop has to store protocol runtime names into
cinterop klibs.

As a first step, this commit adds such names to the native index,
using the new `binaryName` property of `interop.indexer.ObjCProtocol`.

The binary (=runtime) name can be different from the name used in the
source code if `objc_runtime_name` attribute is used.
But this commit doesn't implement reading `objc_runtime_name` in indexer
and postpones that to KT-82200.
So it only fills the binary name if the protocol doesn't have this
attribute and puts `null` otherwise.

^KT-79741