[K/N] Make cinterop set `binaryName` for protocols when instructed
KT-79741 aims to provide an alternative, "direct" implementation for
type checks against Objective-C protocols in the compiler.
To achieve that, cinterop has to store protocol runtime names into
cinterop klibs.

This commit does that: by using information from the indexer, it sets
the `binaryName` parameter of `@ExternalObjCClass` annotation that is
added to all protocols.
(The parameter existed before this commit as well but was used only for
Objective-C classes).

In addition to setting `binaryName`, this commit does another important
relevant change.
`@ExternalObjCClass` has two parameters -- `protocolGetter` and
`binaryName`, and both are used to make the compiler able to generate a
type check against this protocol.

This commit not only sets `binaryName`, but makes the generation of
`binaryName` and `protocolGetter` conditional, depending on the
`-Xccall-mode`.

Basically, this change piggybacks the `-Xccall-mode` machinery,
following its general meaning:
- `indirect` mode:
  allow only the legacy access method that relies on bitcode in the
  cinterop klib (here: `protocolGetter`).
  It also matches the behaviour before the introduction of the
  `-Xccall-mode` machinery, which roughly corresponds to
  `-Xccall-mode indirect`.
- `direct` mode:
  allow only the experimental access method that doesn't need the
  bitcode (here: `binaryName`).
- `both` mode: allow both, let the compiler decide.

^KT-79741