[FIR] Enhance mutability for warning from @Unmodifiable(View)
Mutability enhancement for warning is implemented analogously to
enhancement of nullability for warning.
We support enhancing mutability and nullability at the same time with
either being for warning.
A new warning RECEIVER_MUTABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS is
implemented that is reported when calling a mutating method on a
collection enhanced to readonly for warning.
Finally, support for org.jetbrains.annotations.Unmodifiable and
org.jetbrains.annotations.UnmodifiableView is implemented.

mutabilityAndNullabilityForWarning.kt needed an adjustment because
`@org.jetbrains.annotations:warn` now affects both nullability and
mutability annotations from that package.

#KT-79656 Fixed