[PSI] Fix 'KtDotQualifiedExpression.receiverExpression' for stubs
Sometimes, only one part of the 'KtDotQualifiedExpression'
is serialized to a stub. In cases when the receiver part is omitted,
'receiverExpression' worked incorrectly. In KT-81928 in particular, the
left part was an integer constant that previously was rejected by
'KtDotQualifiedExpression.getChildExpressionsByStub()'.

The commit removes the synthetic limitation for the children stubs
and also fixes the generic issue with `receiverExpression`.

The bug appeared in Standalone after its migration to the stub
infrastructure (in b402f6068e6b70ca566d2bd6babcfa818ebf8f35). The test
is implemented in the Standalone as usual Analysis API tests operate on
the PSI tree.

To avoid similar issues, we should build the test suite that would check
consistency of tree/stub PSI implementations (see KT-82189).

^KT-81928 Fixed