[K/N] Sanitize inferred bundleId.
When producing ObjC Framework, the resulting bundle must have
and identifier `bundleId`. When not overridden with the corresponding
binary option, the identifier is inferred from the output file name
and package names.

According to the documentation
https://developer.apple.com/documentation/bundleresources/information-property-list/cfbundleidentifier#Discussion
`bundleId` must contain only alphanumerics, `-` and `.`. The inferred
identifier did not enforce this rule.

This commit santizes the inferred identifier by replacing all
invalid characters with `-`. The manually provided identifier
(i.e. from the binary option) is not affected.

^KT-81906