[K/N] Work around #KT-82040 during autoboxing transformation
This basically works around having an old library (compiled with K1) in one of
the dependencies. Then #KT-64069 might be in effect which leads to a wrong type
inferred in certain places forcing the autoboxing tranformer to insert an incorrect
cast which then goes off at runtime.

This commit works around this by not inserting these casts (to Nothing) for now
(until K1/1.9.20 is deprecated).

 #KT-82040 Fixed