[JVM] generate bridges for abstract overrides in interfaces
^KT-82467 Fixed

Since Java 8, javac generates bridges for overrides with signature
changes not only in the implementing classes, but also in the
interfaces as default methods.
Before this commit, Kotlin generated bridges in interfaces only for
non-abstract overrides. While it was not an error, now the bridges can
be generated for abstract overrides both in interfaces and classes,
similarly to Java, for the sake of better JVM and binary compatibility.
The behavior can be controlled by the new `-Xinterface-bridges` option,
which is currently set to 'false' by default, but is going to be changed
to a default 'true' value soon.