[Native] KT-82253: Fix macabi experimental support
The end goal of this commit is to fix macabi support for Intel-based macs.
Currently, it is broken because wrong `targetCpuFeatures` are passed.

Effectively, we fix it by changing combination
`iosArm64 + -Xbinary=macabi=x64/arm64`
to
`iosSimulatorArm64 / iosX64 + -Xbinary=macabi=true/false`.

Using iOS Simulator target is motivated by:
- Simpler implementation: we base Intel-based Catalyst support on
Intel-based target. This makes code a bit more robust.
- Simpler KGP integration which does not allow having 2 targets of same
type in a single project.

An alternative is to implement support of macabi via macOS target.
However, it is a bit more complicated solution because it requires
altering the set of platform libraries and making some of them opt-in.
For now, we prefer to avoid doing so because macabi support is
experimental and likely to remain so.