Reflection: fix container of type parameters of local delegated properties
... in case of descriptor-based implementation. In the new
implementation, it works correctly since 7375d523ca.

There is no `containerSource` in the descriptor in case of a local
delegated property, so we could not load the container of its generic
parameters. The solution is to create a fake container source object,
and extract the containing class/package where the container is
computed.

 #KT-77312 Fixed