[FIR] Set proper type arguments for qualified accesses in completion
The UBV checker fails to detect the bound violation, because when it
reads `typeArguments` of the constructor call, it sees `AlphaKey` as
the second argument, not `BetaKey` despite it's written explicitly.
This happens because `typeArguments` are replaced inside
prepareQualifiedTransform` with substitutions from the new fresh type
variables via the `finalSubstitutor`.

These changes are fine as this is already red code, and in general
these `CANNOT_INFER_PARAMETER_TYPE` are strange: they complain about the
type parameter of `checkExactType`, but we do, in fact, have an explicit
argument for it.

^KTLC-373
^KT-80588

# Conflicts:
#	compiler/fir/analysis-tests/testData/resolve/lambdaArgInScopeFunction.fir.txt
#	compiler/fir/analysis-tests/testData/resolve/lambdaPropertyTypeInference.fir.txt