Reflection: remove laziness when computing type parameter container
This only affects old descriptor-based implementation.

Computing the container from a type parameter descriptor is actually
fairly quick because in most cases, it only involves unwrapping things
which are already computed. The only exception is built-in classes (see
`else` in `ClassDescriptor.toJavaClass`), but that is a rare case. And
the container will likely be computed anyway if you do almost any
non-trivial operation with the type parameter, for example call
`equals`.

Note that this only affects type parameters which are obtained from
descriptor-based types, which are at this point still used for
callables. All other type parameter instances already have their
container computed at the point of creation.

The motivation here is to simplify code to prepare for the support of
KT-39661, where `container` will be unified between `KTypeParameterImpl`
(kotlin-reflect implementation) and `TypeParameterReference`
(kotlin-stdlib implementation).