[Stdlib] Introduce `ExperimentalCollectionLiterals` annotation
In this commit, all relevant declarations (namely, companion objects of
arrays, `List`, `Set`, and `Sequence` as well as operators `of` declared
in them) are marked with `SinceKotlin` and this annotation.

Note that `ExperimentalCollectionLiterals` is marked with
`@HidesFromObjC`. See KT-82086: for Swift, special interop must be
implemented, and while the feature is experimental, it's decided to
mute both ObjC and Swift export for operators `of`.

^KT-81722