Native: Rework the API for reading properties in DefFile
All the properties that the C-interop tool can recognize in a *.def
file can be divided into three groups by their type and parser:
- string list properties
- nullable string properties
- boolean properties

With this commit, we are introducing a more formalized approach for
reading and parsing of such properties.

Also, all known *.def file properties are assembled in a single place:
`sealed interface DefFileProperty` with the three `enum class`
members corresponding to the three property groups.

^KT-81411