Reflection: fix javaType for types created with withNullability
"Reset" `computeJavaType` when we're copying a type with a new
nullability. Also do the same if we're applying DNN to a type, which is
used in `KClass.allSupertypes` as well as type equality, but I couldn't
create a test where this behavior change would be visible.

For descriptor-less types (`SimpleKType`), Java type has already been
correctly reset when invoking `withNullability`. However, for them we
need to adjust the classifier.

 #KT-82350 Fixed