FIR, REPL: introduce mechanism for search scopes use with FIR sessions
to be able to filter out the deserialized declaration by the search
scope of the particular module data and session.
This allows constructing the correct REPL sessions structure in case
of adding dependencies to the subsequent snippets.
All other uses of this mechanism are hidden or disabled by default
implementations.
The change requires all known and used in the search scopes filesystems
to be threaded to the `VfsBasedProjectEnvironment`, and therefore
relevant changes at the use sites.

#KT-82369 fixed