[FIR] Java: cover case with binary annotation which resolves to source
Even Java resolve might in exceptional cases return a light
class for a source Kotlin class. `JavaClassImpl` forbids
usages of light classes, so this API cannot be used as is
in this case.
Since normally such situations are impossible (this setup
should lead to a runtime exception due to a missed library),
it should be fine to just ignore such cases instead of trying
to work with source declarations

^KT-82076 Fixed