[Analysis API] Introduce API for finding KDoc for KaDeclarationSymbol and KtDeclaration
KT-63339

Some Analysis API users (IDE, Dokka) need to access KDoc's of the
symbols. Previously, each user implemented its own wrapper functions on
top of AA. This commit provides a unified API for accessing KDoc.
Implementation is based on the current implementation of the same
functionality in IDE, as it is considered a 'ground truth.'


Merge-request: KT-MR-24050
Merged-by: Azat Abdullin <Azat.Abdullin@jetbrains.com>