[K/N][tests] Ignore objCReferenceGlobals.kt when `gcType=NOOP`
5608978 introduces a new test for the new feature --
`objCReferenceGlobals.kt` checking that Objective-C reference globals
can be properly accessed when the cinterop klib is produced with
`-Xccall-mode direct`.
This test fails when run with `gcType=NOOP`.

The test heavily relies on the GC doing its job: it checks that
reading and writing the global respects Objective-C reference counting
by running Kotlin GC and checking that the Objective-C objects held by 
Kotlin objects are deallocated as expected.

Without a GC in place, it can't reasonably check this. So it makes sense
to simply ignore the test in this mode to make it green.

This commit does exactly that.