[IR] Remove some special handling for trim functions in the interpreter
There was some special handling for trimMargin and trimIndent however,
since we now also use the map to determine which functions are
interpretable we had to add them to the map.
This enabled the removal of some of the special cases from the
IntriniscImplementation with only the trimMargin with the default
argument remaining. Which at the moment cannot be interpreted because
the expression of the default argument is an ErrorExpression and crashes
the interpreter.

^KT-81660