[KMP Separate Compilation] Check parameter names when actualizing ObjC methods
`platform.Foundation.NsCalendar` contains two `getEra` overloads with
the same signatures except parameter names. This is legal in ObjC, so
we support it.

To prevent an ambiguity during actualization of `NsCalendar`, we make
consider methods annotated by @ObjCMethod as having stable parameter
names and we add parameter names checking to
AbstractExpectActualMatcher.

#KT-82418 Fixed