check if new constraint is useless against all previous constraints
Currently, 'addConstraint' will check if the added constraint is
"useless" (i.e., adding an UPPER constraint when EQUAL exists). There
are also special condition where a constraint is never considered
useless, for example when it replaces a constraint from an upper bound.

In the flow of 'addConstraint', if the constraint is not deemed useless,
the next step is checking for whether the previous constraint can be
simplified. If yes, the function immediately returns the simplified
constraint.

Normally, the method would iterate through the list of constraints and
find one that makes the new constraint useless or can be simplified by
it.

But when the first constraint in the list happens to be from an upper
bound, and it can be simplified, then the simplified constraint will
be added to the END of the list. If the same constraint is added again,
instead of skipping it becuase it would be useless, only the first
constraint is checked, and the new constraint is added twice.

This commit instead goes through all constraints first to check if the
constraint is made useless by any of them, and only then tries to
simplify.