[IR][K/N] Split tests processing and initialization
This is a preparation step for #KT-74223. TestProcessor adds a bunch of top-level fields,
all having the same name, which is fine for K/N backend but triggers a validation during IR
serialization. These top-level fields serve as initializers for all tests within a file:
their initializers create all necessary tests; the fields are marked with @EagerInitialization
forcing the backend to run their initializers at start-up.
This commit goes around this by putting the initialization code into a special function instead
of a top-level field. During the second part these functions are replaced with fields as was before.
The idea here is that the first part can be done during the first stage, making it possible to
serialize IR to a klib; while the second part can be performed during final binary emitting.

 #KT-74223