[K/N] remove `fullCompilerConfiguration` from `metadata.properties`
The file in question stores metadata for the Kotlin/Native caches.
The `fullCompilerConfiguration` doesn't seem to serve any functional
purpose. Originally was added to make debugging easier (by seeing how
caches were built).

Unfortunately the configuration contains path strings, which leads to
the distribution being not reproducible. It is possible to omit only
the path-related propertiesm, but the blacklist approach doesn't prevent
from those creeping in later on. Besides a deeper look showed that some
of the values do not have a proper  - so those need to be
omitted as well for the sake of reproducibility

Since the original purpose - debugging - doesn't seem to be used often,
it seems to be easier to remove this information alltogether at this time.

If there is a need to re-introduce some debuggin info in the future,
it will need to be done with reproducibility in mind.