[BTA] Add custom compiler plugins representation decomposition
However, there's no clear way to do the opposite transformation from unstructured way (applied via `applyArgumentStrings`) to the structured. In follow-up commits, there will be some marker indicating that compiler plugins are configured, but they cannot be retrieved through the BTA custom representation. In this case, the arguments applied via `applyArgumentStrings` should be applied 'as is' combined with the plugins configured via the structured way. This can in some cases cause incompatibility as there're 2 approaches to pass compiler plugins to the compiler, -Xplugin and -Xcompiler-plugin which cannot be combined together.
^KT-80681 In Progress