[LL] forbid redeclaration for library sources
Library sources may have equivalent declarations in the same module,
so they have to be filtered via `ConeCallConflictResolver`.
In particular, `ConeEquivalentCallConflictResolver` is supposed to
deduplicate `expect` and `actual` declarations.
The issue is covered by
`StdLibSourcesLazyDeclarationResolveTestGenerated.testWrappedInt` test.
The flaky behavior of this test was caused by the non-stable order of
`kotlin.toString` expect and actual declarations inside the stdlib
sources.
As a result, in one case
`ConeEquivalentCallConflictResolver.filterOutEquivalentCalls` was
enough to drop `expect` declarations since it was in the first position,
and we also had another declaration from the builtins module.
If the `actual` variant was the first one, then this condition wasn't
reached, and the result contained both `expect` and `actual` since
they weren't treated as equivalent.
13021e43d42d60f6e85fcb4e25109647ced9e998 commit was the place where
the order of `expect` and `actual` became a factor

^KT-76487 Fixed