Make toList() safer on Collections with size 1
It's possible for library code to define a Collection whose size
property is one, but producing an iterator without any next element.
This is probably some kind of mistake, but it's also very hard for a
user of that library to deal with the error that toList() generates
under those circumstances.  If size is 1, conditionally generate an
empty list if the initial iterator has no next element, consistently
with other collection snapshotters.