Native: Allow substitution only for the limited set of properties
We do not want property substitution to be possible for all manifest
file properties, because this can be the source of hardly debuggable
problems. Instead, we would like to permit substitution only for
a limited set of properties where it is known to be used in practice
and is known to be safe.

^KT-81411