Reflection: use delegated properties in KProperty implementations
This is a refactoring that should not affect any user-visible behavior.

Originally, this was done to prevent an initialization loop: during
static initialization of DescriptorKProperty, we would need to create
KProperty objects for delegated properties' getValue, but those objects
were eagerly wrapped into DescriptorKProperty. After KT-81974, this
initialization loop is no longer possible.