[Wasm] Reimplement Wasm compilation phase in two phases
It makes Kotlin/Wasm backend be able to compile code in two steps - first it compile the code into WasmIR from BackendIR and then link and compile wasm file. This refactoring makes possible for Wasm to release BackendIR before linking and emitting the binary. That gives us a possibility to consume less memory in a moment.

As a part of KT-82126