[Doc] Describe test recommendations in the "Code authoring guideline"
The "Code authoring guideline" hints that changes in code should be
covered with automated tests but never mentions this explicitly as the
responsibility of code authors.

This commit expands the "Code authoring" section of the document to
mention that an MR should be well-covered with automated tests and gives
some explanations and recommendations on that topic.