FE/fixation: optimize readiness calculation in "forbidden" case
In this commit, we made calculation faster in case when variable fixation
is forbidden. This optimization changes nothing semantically,
as with "forbidden" flag we anyway can't select this variable for fixation,
so differences between "forbidden" variables in fact have no influence.
Changes are suggested by Denis Zharkov.

Related to KT-77939