[K/N] Don't access an allocated Objective-C class until it is registered
96964e3 introduces a problem -- Objective-C interop runtime crashes with
errors like

    objc[6646]: no class for metaclass 0x1033e43c0

when the thread state checker is enabled (e.g. running tests with
`-Pkn.useThreadStateChecker=true`)

The root cause is the following: that commit makes
`Kotlin_ObjCExport_createTypeInfoWithKotlinFieldsFrom` call an
Objective-C method on the newly allocated class before it is properly
registered with `objc_registerClassPair`.

The new call is within `IsKotlinObjCClass` called by `createTypeInfo`.
Surprisingly, the call itself worked fine. But enabling the thread state
checker also generates
`class_getMethodImplementation(object_getClass(cls), ...)`,
which in turn crashes.

This commit fixes that by moving the creation of the `TypeInfo` after
`objc_registerClassPair`.
It is safe, as the created `TypeInfo` is not supposed to be used until
the class is published later.

^KT-82669 Fixed