[PL] Change the behavior of getting name of removed function in K/Native
The new behavior is to return the name instead of throwing a linkage error.
It is done by removing the checks for `reflectionTargetLinkageError` and
adding a name property to the `LinkageError` subclass of `KFunctionDescription`.
Thanks to that `KFunctionImpl.name` works regardless of whether we have a linkage error or not.

It is possible to add the name property to the `LinkageError` subclass thanks to the
 `IdSignature.guessName` function called during stub generation.

^KT-78365