[FIR] Reimplement completion for old resolution of collection literals
In this commit effectively two things are done. First, the ways we treat
array literals in cases when they are in the correct/wrong positions are
unified. That way, we don't miss not fully resolved atoms for lambdas
and callable references when they are inside of incorrectly used array
literal.

Second, for empty collection literal when there's no good (i.e., array)
expected type, we resolve it with `Array<Any>` expected type now.
That way, `arrayOf` is always resolved and then transformed back to
collection literal.

^KT-82684 Fixed
^KT-81777 Fixed
^KT-76150 Fixed

RCA: (for KT-82684) in the scope of KT-77361 a branch
`is NotEnoughInformationForTypeParameter<*>` in
`ConstraintSystemError.mapConstraintSystemError` was shortened with only
the subcase of delegated constructor call left. The idea was that the
diagnostic was duplicated by `CANNOT_INFER_PARAMETER_TYPE`. However,
in KT-82684 `NEW_INFERENCE_NO_INFORMATION_FOR_PARAMETER` was the only
diagnostic being reported (due to how `FirArrayOfCallTransformer` is
implemented -- normally, for synthetic calls there at least would be
some `ConeErrorType` in the return type, but here we build error
expression instead).