[K/N][tests] Normalize build directory in compiler output
There are K/N compiler tests that check compiler output. Typically, they
normalize the output to compare it against expected static outputs.
For example, this involves replacing paths to the test data directory
with the `$TESTDATA_DIR$` placeholder.

This commit makes this normalization also replace the test build
directory path with `$BUILD_DIR$`.

This simplifies testing the compiler output when it contains paths to
built artifacts.

^KT-81828