Compiler, reflection: fix `isExternal` for native Java methods
`isExternal` modifier was loaded incorrectly for native Java methods,
both in K1 and K2. It did not affect anything in the compiler, but was
visible in the descriptor-based implementation of kotlin-reflect via
`KFunction.isExternal`.

The change in FakeOverrideBuilderStrategy is needed because otherwise
`isExternal` modifier for fake overrides was always erased on JVM. Now
this code only erases `isExternal` if the original declaration's
containing class was external, which is only possible on JS, where this
code is supposed to work.

 #KT-82699 Fixed