[FIR] Reuse cached implicit receiver scope for explicit receivers
Same as last commit, this fixes some false positive type mismatches
because when we recreate the scope, we capture projections and create
captured types that don't compare as equal.

Some UNRESOLVED_REFERENCE diagnostics in combination with
INACCESSIBLE_OUTER_CLASS disappeared because previously, we would have
an `FirThisReceiverExpression` with an error type and when we created
a scope from its return type, it would be empty. Now, we reuse the
implicit receiver that has a proper return type and thus has a proper
scope.

#KT-68606 Fixed
#KT-71024