[K/N] set useXcode to solve a SecurityManager issue
There is a system in our repository which complains if a test uses
unexpected files - see
`repo/gradle-build-conventions/project-tests-convention/Readme.md`

This system needs to be told that a test uses XCode, otherwise the test
fails with a AccessControlException upon access to it. This only happens
when `KONAN_USE_INTERNAL_SERVER`  is enabled.