[wasm]: Relax js(...) call constraints somewhat when lowering.
We have many constraints on how js(...) can be used when targeting the
wasm backend, in particular that functions using it must only call
js(...) and do nothing else. The lowering phase assumed this, which
was mostly fine because there are also syntactic validation checks to
make sure uses of js(...) conformed to these restrictions.

However, function annotations such as @Composable can actually do
arbitrary code transformations which break the constraints that the
lowering pass assumes, even if the original function passes the
validation steps pre-transformation. In order to not have to maintain
a whitelist or blacklist of allowable code annotations, this change
attempts to relax the restrictions just enough to allow code
transformations which preserve the semantics of the js call (which
should be most well-behaving annotations) to work correctly.

In particular, the lowering pass now supports multiple calls to js in
unrestricted positions in a function which may do other things as
well, for example in conditional branches or other control flow
constructs that a code transformation may wrap around the original
code. However, the restrictions which remain create slightly
unintuitive semantics which makes it unsuitable to expose these
relaxed semantics to user code, so none of the user-code validation
checks have been lifted. In particular, the remaining restrictions
which remain on the lowering side due to implementation reasons are:

* Only function parameters can be referenced in js code.
  ** Closing-over local variables, while possible, essentially force a
     shared-variables or closure conversion type analysis to occur,
     and for assignment to work, heap allocation of value cells is
     entailed.
* The return type of all calls to js(...) is taken directly from the
  return type of the entire function.
  ** While it may make more sense to try and figure out the correct
  expected return type from the calling context, the type system
  doesn't work that way.

Since js(...) calls are opaque to code transformation annotations,
it's not really possible that code annotations which seek to preserve
the semantics of the js call will change the expected return type or
introduce locals that the js call needs to reference. For user code,
the return type quirk is going to be too unexpected.

No functional difference should otherwise be observed.

Fixes user program in ^KT-79411.