[IR] Improve the `copyValueAndTypeParametersFrom` helper
This IR util remapped the new value parameters using the new type
parameters, but didn't remap the return type of the target function,
which is just as essential. Naturally, nearly on each its call site
we would forget to substitute the return type using the new type
parameters.

This is fixed here, the function now uses the same substitution map
to not only remap types in value parameters, but also in the return
type of the target function.

To reflect the fact that the function is no longer only about
_parameters_, it has been renamed.