FIR2IR: Postpone rarely necessary subtyping at applySuspendConversionIfNeeded
The idea is that `isSuspendOrKSuspendFunctionType` is relatively cheap 
O(1) operation, and rarely returns true as many of expected types 
are not function ones.

It improves Translation-phase performance around 5%, and the overall
pipeline around 0.5%
For me, it was quite surprising that `isSubtypeOf` might be notably
expensive, but on the other hand, it's being called almost for all
expressions.

^KT-82757 Fixed