[K/N] Introduce GlobalDecl.directAccess encapsulating .binaryName
Currently, a `GlobalDecl` in interop Indexer is considered compatible
with the `-Xccall-mode direct` mode <=> it has a non-null `.binaryName`.

This commit refactors this to encapsulate the binary name if compatible
or the incompatibility reason otherwise, by introducing the new sealed
class `DirectAccess` and storing an instance to `GlobalDecl`.
The incompatibility reason is used when generating the actual Kotlin
declarations to make it visible to users.

This is a refactoring. It doesn't change any observed behavior but
serves merely as a preparation for future changes.

This commit also adds unit tests for the value of the new property.
Some of them are to be changed within KT-83039.

As a side effect, the changed code doesn't use the unconventional term
"binary name" anymore. The conventional term "symbol" is used instead.

^KT-79755