[Wasm] Fix test runner in multimodule mode
As a part of KT-76275

Now Wasm backend generates startUnitTest function (and exports it for the user).
startUnitTest registers a bunch of module's test and than run them.
But if startUnitTest run's twice than double registration occurs, so tests will be runned twice.
So we need to clear registrations after each run.