[K/N] toggle isNative property in all nativeTestTask's
The `test-inputs-check` restricts access to files beyond
the task inputs. For Kotlin/Native there is a special whitelist,
which includes xcode and other stuff vital for the tests.

Toggling the `isNative` property for each Kotlin/Native test
is cumbersome. This commit sets `isNative` to `true` for
each `nativeTestTask` automatically whenever the `test-inputs-check`
is present.