[K/JVM] Copy all attributes in JvmUpgradeCallableReferences
Unfortunately, a lot of code relies on the fact that the localClassType
attribute is the same for nodes with same attributeOwnerId and different
for nodes with different attributeOwnerId.

Coping only this attribute breaks this invariant, as if two classes
had same attributeOwnerId, they would have same localClassType, but no
longer connected.

In particular, this can happen in for function references or lambdas
which were copied for some reason before being transformed to classes.
This is not happening in compiler, but can happen in plugins, e.g.
kotlinx.serialization does this for serializable classes default values.

In general, it's not completely correct to make them same,
as it can lead to deduplication of potentially differently lowered
classes. But in practice, this shouldn't be the case (as having
connected references is rare, and they should be lowered to equivalent
classes in that case).

This probably can be removed after moving upgrade
lowering earlier in pipeline.

#KT-82788 Fixed