K2: use completion in resolve procedure for equality operator calls
Before this commit, we resolved a right argument of == in
ExpectedType(Any?) resolution mode. In some corner cases like KT-81763,
we in fact need more precise knowledge here to resolve all types properly.
So in this commit, we started to resolve the right argument in
dependent resolution mode and added a completion call at the end.
This commit also adds certain changes in CFA/DFA, as FirEqualityOperatorCall
is now a call that follows almost normal call resolve procedure.

#KT-81763 Fixed