[LL] FirDesignation: rework fallback class searcher
Typically, the designation search goes via `FirElementFinder`, but
it can be used if the target declaration is present in the FirFile tree.
For other cases, the provider-based search was used. However, it is not
reliable since many things can go wrong, so no one guarantees that the
result class would be the one you are searching for.
To fix this, `LLContainingClassCalculator` (which is PSI-based) is used
to properly search for containing classes – this solves all possible
ambiguity (like `FirElementFinder`).
`resolveChunk` still remains because there is no sure that all cases
could be covered by `LLContainingClassCalculator`

^KT-82948 Fixed