[K/N] Introduce CString for malloc-allocated C strings in libclangext
libclangext is our extension to libclang. It contains functions that
expose certain functionality from Clang to cinterop indexer when this
is not available directly in libclang.

To achieve that, libclangext duplicates some code from libclang and has
some expectations over libclang's internal structures, to be able to
accept values produced by libclang. This is fragile. This commit
improves one part of this.

Some functions in libclangext need to return dynamically allocated
strings (e.g. created with strdup). For example,
`clang_Cursor_getSwiftName`. Another string-returning function will need
to be introduced for KT-82200.
Those dynamically-allocated strings need to be somehow represented in
libclangext.

Technically, libclang already has the relevant type CXString, but
reusing it would require libclangext to break libclang's encapsulation
even more.

Another option for the representation is to use `char*` as the return
type and expect the callers to dispose it with `clang_free`.
There are two problems with this approach:
* That expectation is too implicit.
* `clang_free` is not guaranteed to call libc's `free`, so this again
  relies on libclang implementation.

This commit implements another option: it introduces the new type
`CString`, and a dedicated function to dispose it that calls `free`
directly.

^KT-82200