[K/N] Add clang_Cursor_getObjCProtocolRuntimeName to libclangext
To fix KT-82200, the cinterop indexer needs to retrieve runtime names
of Objective-C protocols when they are specified with the
`objc_runtime_name` attribute.

The problem is that libclang doesn't provide much tools to do that.
As a workaround, this commit adds a new function,
`clang_Cursor_getObjCProtocolRuntimeName`, to our extension to libclang,
named libclangext.

As other things in libclangext, this function is implemented using C++
API for Clang AST, which naturally provides a simple way to retrieve
the runtime name.

^KT-82200