[wasm]: Don't export Kotlin call helpers for JS interop. ^KT-82843
Instead of exporting trampolines which
__convertKotlinClosureToJsClosure will use to call Kotlin closures, we
pass references to them as arguments instead. Care needs to be taken
to forward-declare the references.

IrRawFunctionReference is reused for the purpose of representing this
kind of raw functional reference at the WebAssembly level. The DCE
code has been changed to not special-case on the trampolines anymore
and instead has been generalized to work with
IrRawFunctionReferences. The net effect is that we remove special-case
code for trampolines and have laid the groundwork for representing
more stuff in the IR explicitly with RawFunctionReferences.

Additionally, the unlocks the possibility of defining converter
functions once per arity, instead of once per type signature.