[FIR] Report `CANNOT_INFER_PARAMETER_TYPE` on transformed `arrayOf`s
^KT-82684 Fixed

RCA: in the scope of KT-77361 a branch
`is NotEnoughInformationForTypeParameter<*>` in
`ConstraintSystemError.mapConstraintSystemError` was shortened with only
the subcase of delegated constructor call left. The idea was that the
diagnostic was always already reported. However, in KT-82684
NEW_INFERENCE_NO_INFORMATION_FOR_PARAMETER was the only diagnostic being
reported (due to how `FirArrayOfCallTransformer` is implemented --
it creates `FirErrorExpression` wrapper for array literal in case the
call was resolved with errors).